[refactor/#330] AlertModal 폴더명 수정, 정보수정 complete 모달을 AlertModal 폴더로 이동
